+++ lib/amazonka/src/Amazonka/EC2/Metadata.hs
@@ -217,9 +217,9 @@ data Metadata
   | -- | Any other piece of metadata specified by arbitrary key.
     --
     -- This is provided for forward compatibility: should there ever appear any gap
-    -- between a future version of AWS API and what (possibly an outdated version of)
-    -- this library covers -- the constructor gives the option to work around.
-    XtendedMeta !Text
+    -- between a future version of IMDS API and what this library covers,
+    -- this constructor provides a workaround.
+    Other !Text
   deriving stock (Eq, Ord, Show, Generic)
 
 instance ToText Metadata where
@@ -258,7 +258,7 @@ instance ToText Metadata where
       Services m -> "services/" <> toText m
       Spot m -> "spot/" <> toText m
       Tags m -> "tags/" <> toText m
-      XtendedMeta m -> m
+      Other m -> m
 
 -- | Metadata keys for @autoscaling/*@.
 data Autoscaling